Revised Fullscreen Code + Touch Controls
Fullscreen code didn't work on PS4 and had issues with resizing. I updated the code with status values for the height for the nav bar and the video controls.

Added Touch controls. Double tap on the left side of the screen to seek back 10, right to advance 10, and the center to hide the controls.